@@ -60,9 +60,9 @@
     "path": "^0.12.7",
     "promise-polyfill": "^8.1.3",
     "random-hash": "^4.0.1",
-    "react": "^16.13.1",
+    "react": "^18.0.0",
     "react-contenteditable": "^3.3.6",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.5.0",
     "react-transition-group": "^4.4.1",
     "styled-components": "^5.2.0",
